requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
coverage==7.9.2
Django==4.2.23
django-allauth==65.11.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.3.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dotenv==1.1.1
pytz==2025.2
requests==2.31.0
resend==0.8.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
